backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
aiomysql = "^0.2.0"
cryptography = "^42.0.5"
mysql-connector-python = "^8.4.0"
google-cloud-storage = "^2.16.0"
alembic = "^1.13.1"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"
coverage = "^7.5.0"
ruff = "^0.4.3"
mypy = "^1.10.0"


[tool.ruff]
exclude = [
    ".ruff_cache"
]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "F", # Flake8
    "B", # Black    
    "E", # error
    "W", # warning
    "I", # isort
]
ignore = []
fixable = ["ALL"]


[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
